moved configureation files to config directory
